refactor : 패키지 이동 view 패키지로 InputView와 OutputView를 함께 패키지로 묶는다.
