+++ src/main/java/bridge/controller/BridgeController.java
@@ -1,6 +1,8 @@
 package bridge.controller;
 
 import bridge.*;
+import bridge.view.InputView;
+import bridge.view.OutputView;
 
 import java.util.List;
